requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
click==8.1.3
Django==4.2
djangorestframework==3.14.0
pytz==2023.3
redis==4.5.4
rq==1.13.0
sqlparse==0.4.4
